requirements.txt
prometheus-client==0.20.0
requests==2.32.3
pyyaml==6.0.2
